feat(publish): support multiple providers
Allow publishing to e.g. NPM, GitHub, and the App Hub simultaneously.
