--- plugins/publish-app-hub.js
@@ -1,10 +1,0 @@
-const appHub = require('../custom/semantic-release-app-hub.js')
-
-exports.appHubPlugin = ({ pkgRoot, baseUrl, channel }) => [
-    appHub,
-    {
-        pkgRoot,
-        baseUrl,
-        channel,
-    },
-]
